Remove test_pauschbetrag_steps
